--- app/assets/components/MyMap2.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-export default class MyMap2 extends Component {
-
-  render {
-    return (
-      <div></div>
-
-    )
-  }
-
-
-}
