+++ app/index.js
@@ -4,5 +4,5 @@ import { render } from 'react-dom'
 import App from '../app/assets/components/App'
 
 
-render(<App initialCenter={ {lng: -90.1056957, lat: 29.9717272} } />,
+render(<App />,
   document.getElementById('main'))
